move sealer cert command to cluster module
