+++ cmd/sealer/cmd/alpha/alpha.go
@@ -35,6 +35,5 @@ func NewCmdAlpha() *cobra.Command {
 	cmd.AddCommand(NewMergeCmd())
 	cmd.AddCommand(NewUpgradeCmd())
 	cmd.AddCommand(NewGenCmd())
-	cmd.AddCommand(NewCertCmd())
 	return cmd
 }
